pyproject.toml
[project]
name = "umama-fan-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imageio-ffmpeg==0.6.0",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "tqdm>=4.67.3",
    "winrt-runtime>=3.2.1",
    "winrt-windows-foundation>=3.2.1",
    "winrt-windows-graphics-imaging>=3.2.1",
    "winrt-windows-media-ocr>=3.2.1",
    "winrt-windows-storage>=3.2.1",
    "winrt-windows-storage-streams>=3.2.1",
]
